+++ pgrpc_a_example.py
@@ -576,7 +576,7 @@ async def do_unstake(client: SuiGrpcClient):
 
 
 async def main():
-    """."""
+    """Example main."""
     client_init: SuiGrpcClient = None
     try:
         client_init = SuiGrpcClient(
@@ -585,6 +585,8 @@ async def main():
                 profile_name="devnet",
                 # profile_name="testnet",
                 # profile_name="mainnet",
+                # profile_name="test-arch",
+                # profile_name="main-arch",
             ),
         )
         print(f"Active chain profile   '{client_init.config.active_profile}'")