+++ pysui/sui/sui_grpc/suimsgs/google/rpc/__init__.py
@@ -11,7 +11,7 @@ import betterproto2
 
 from ...message_pool import default_message_pool
 
-_COMPILER_VERSION = "0.7.1"
+_COMPILER_VERSION = "0.8.0"
 betterproto2.check_compiler_version(_COMPILER_VERSION)